Fix single date selection in organization audit logs
